+++ v4/localenv_test.go
@@ -1158,6 +1158,141 @@ func TestLocalEnvLookupHomeDirReturnsFalseIfUserDoesNotExist(t *testing.T) {
 //
 // ----------------------------------------------------------------
 
+func TestLocalEnvReplacePositionalParamsSetsThePositionalParams(t *testing.T) {
+
+	// ----------------------------------------------------------------
+	// setup your test
+
+	testData := []string{
+		"one",
+		"two",
+		"three",
+		"four",
+		"five",
+		"six",
+		"seven",
+		"eight",
+		"nine",
+		"ten",
+	}
+
+	env := NewLocalEnv()
+
+	expectedLen := 10
+	expectedResult := []string{
+		"$#=10",
+		"$1=one",
+		"$2=two",
+		"$3=three",
+		"$4=four",
+		"$5=five",
+		"$6=six",
+		"$7=seven",
+		"$8=eight",
+		"$9=nine",
+		"$10=ten",
+	}
+
+	// ----------------------------------------------------------------
+	// perform the change
+
+	actualLen := env.ReplacePositionalParams(testData...)
+	actualResult := env.Environ()
+
+	// ----------------------------------------------------------------
+	// test the results
+
+	assert.Equal(t, expectedLen, actualLen)
+	assert.Equal(t, expectedResult, actualResult)
+}
+
+func TestLocalEnvReplacePositionalParamsReplacesAllExistingPositionalParams(t *testing.T) {
+
+	// ----------------------------------------------------------------
+	// setup your test
+
+	seedData := []string{
+		"one",
+		"two",
+		"three",
+		"four",
+		"five",
+		"six",
+		"seven",
+		"eight",
+		"nine",
+		"ten",
+	}
+
+	testData := []string{
+		"new one",
+		"new two",
+		"new three",
+	}
+
+	env := NewLocalEnv()
+	env.SetPositionalParams(seedData...)
+
+	expectedLen := 3
+	expectedResult := []string{
+		"$#=3",
+		"$1=new one",
+		"$2=new two",
+		"$3=new three",
+	}
+
+	// ----------------------------------------------------------------
+	// perform the change
+
+	actualLen := env.ReplacePositionalParams(testData...)
+	actualResult := env.Environ()
+
+	// ----------------------------------------------------------------
+	// test the results
+
+	assert.Equal(t, expectedLen, actualLen)
+	assert.Equal(t, expectedResult, actualResult)
+}
+
+func TestLocalEnvReplacePositionalParamsUpdatesDollarHash(t *testing.T) {
+
+	// ----------------------------------------------------------------
+	// setup your test
+
+	seedData := []string{
+		"one",
+		"two",
+		"three",
+	}
+
+	testData := []string{
+		"new one",
+		"new two",
+		"new three",
+		"new four",
+		"new five",
+	}
+
+	env := NewLocalEnv()
+	env.SetPositionalParams(seedData...)
+
+	origHash := env.Getenv("$#")
+	assert.Equal(t, "3", origHash)
+
+	expectedResult := "5"
+
+	// ----------------------------------------------------------------
+	// perform the change
+
+	env.ReplacePositionalParams(testData...)
+	actualResult := env.Getenv("$#")
+
+	// ----------------------------------------------------------------
+	// test the results
+
+	assert.Equal(t, expectedResult, actualResult)
+}
+
 func TestLocalEnvSetPositionalParamsSetsThePositionalParams(t *testing.T) {
 
 	// ----------------------------------------------------------------